Precision Reader Files added
See Instructions.txt.
The format of Precision data is different depending upon what has been downloaded. Sometimes, for example, there is a 30second run of AF data from a PentaRay. Other times, there is a large map of thousands of points. This code works for 30s runs but needs further validation for entire maps, due to recent formatting changes of the download. Happy to collaborate on this ...